Fix Climb List initial data problem and add climb name filter
Originally I tried to set the climbs list int he queuecontext from a layout.tsx.
But turns out layout.tsx doesnt have access to the searchparams.
So instead I moved it back to the page.tsx for the list route. And we inject the results as a prop so that the user
doesnt get a loading state.

For other routes using the queue provider this should be fine, as the only other part that wouldhave a loading state are
the queue control buttons int he bottom bar