Only set climb uuid hash when a climb is set active
The auto updating hash while scrolling was a bit yank
and possibly slow. Also the old restore was restoring
based on refs, which has to wait  for the first client
side render. Instead we now use getElementById to get
the col of the climb uuid then we scroll to that, which
is possible on the SSR result