Fix logging of ascents & attempts
Aurora removed all their rest apis and now we have to use webforms, which is a
bit questionable. Not sure why they did this, maybe they think rest apis are fair game
but webforms are not? Either way, this change fixes it again, but cant promise it'll stay
like that for very long.